WIP: Fixing more bugs with angle calculations, flushing out moving trains
Also run in compatability mode. Unfortuntely, it seems that forward+
mode causes suttering; it _seems_ to be something with the rendering
pipeline + maybe the graphics card(NVIDIA 3070). There may be ongoing
bugs in godot. Timeboxed it and couldn't figure it out, so will revisit
later on if I find I need forward+